+++ packages/fe-container/src/__tests__/App.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import {screen, waitForElementToBeRemoved} from '@testing-library/react'
+import {screen} from '@testing-library/react'
 import nock from 'nock'
 import userEvent from '@testing-library/user-event'
 
@@ -22,6 +22,14 @@ describe('App test', () => {
       .persist()
       .get('/api/v1/microlc/configuration')
       .reply(200, {
+        theming: {
+          header: {
+            pageTitle: 'Mia Care',
+            favicon: 'https://www.mia-platform.eu/static/img/favicon/apple-icon-60x60.png'
+          },
+          variables: {},
+          logo: 'https://media-exp1.licdn.com/dms/image/C4D0BAQEf8hJ29mN6Gg/company-logo_200_200/0/1615282397253?e=2159024400&v=beta&t=tQixwAMJ5po8IkukxMyFfeCs-t-zZjyPgDfdy12opvI'
+        },
         plugins: [{
           label: 'entry_1',
           id: '1',
@@ -33,11 +41,14 @@ describe('App test', () => {
         }]
       })
     const toggle = await screen.findByTestId('top-bar-side-menu-toggle')
+    expect(global.window.document.title).toEqual('Mia Care')
     expect(await screen.queryByText('entry_1')).toBeNull()
     userEvent.click(toggle)
-    expect(await screen.findByText('entry_1')).toBeTruthy()
+    const entry1 = await screen.findByText('entry_1')
+    expect(entry1).toBeTruthy()
     userEvent.click(toggle)
-    await waitForElementToBeRemoved(() => screen.queryByText('entry_1'))
+    // @ts-ignore
+    expect(entry1.parentElement.parentElement.classList).toContain('ant-dropdown-hidden')
     mockedResponse.done()
   })
 })
