+++ packages/fe-container/src/colors.less
@@ -1,4 +1,4 @@
 @primary-color: #1DA57A;
 @body-background: #e8e8e8;
-@layout-body-background: rgba(0, 0, 0, 0.25);
+@overlay-background: rgba(0, 0, 0, 0.25);
 @layout-header-background: white;
