+++ packages/fe-container/src/components/layout-content/LayoutContent.less
@@ -2,3 +2,8 @@
   height: 100%;
   width: 100%;
 }
+
+.layout-container-overlay {
+  opacity: 0.2;
+  background-color: black;
+}
