+++ packages/fe-container/src/components/layout-content/LayoutContent.tsx
@@ -2,18 +2,18 @@ import React, {useCallback, useContext} from 'react'
 import {Layout} from 'antd'
 
 import {MenuOpenedContext} from '../../contexts/MenuOpened.context'
-import {AnimatedLayoutSider} from './animated-layout-sider/AnimatedLayoutSider'
 import {LayoutCenter} from './layout-center/LayoutCenter'
 
 import './LayoutContent.less'
 
 export const LayoutContent: React.FC = () => {
   const {isMenuOpened, setMenuOpened} = useContext(MenuOpenedContext)
+
   const closeSideMenu = useCallback(() => setMenuOpened(false), [setMenuOpened])
+
   return (
-    <Layout className="layout-container">
-      <AnimatedLayoutSider isOpened={isMenuOpened}/>
-      <LayoutCenter closeSideMenu={closeSideMenu}/>
+    <Layout className={isMenuOpened ? 'layout-container-overlay' : ''} onClick={closeSideMenu}>
+      <LayoutCenter/>
     </Layout>
   )
 }
