+++ packages/fe-container/src/components/layout-content/layout-center/LayoutCenter.tsx
@@ -1,4 +1,3 @@
-import PropTypes from 'prop-types'
 import React, {useCallback, useContext} from 'react'
 import {ConfigurationContext} from '../../../contexts/Configuration.context'
 import {Plugin} from '@mia-platform/core'
@@ -6,13 +5,7 @@ import {Route, Router, Switch} from 'react-router-dom'
 import {Layout} from 'antd'
 import {history} from '../../../plugins/PluginsLoaderFacade'
 
-const layoutCenterProps = {
-  closeSideMenu: PropTypes.func.isRequired
-}
-
-type LayoutCenterProps = PropTypes.InferProps<typeof layoutCenterProps>
-
-export const LayoutCenter: React.FC<LayoutCenterProps> = ({closeSideMenu}) => {
+export const LayoutCenter: React.FC = () => {
   const configuration = useContext(ConfigurationContext)
   const hasRoute = useCallback((plugin: Plugin) => plugin.pluginRoute, [])
   const routerMapper = useCallback((plugin: Plugin) => (
@@ -22,7 +15,7 @@ export const LayoutCenter: React.FC<LayoutCenterProps> = ({closeSideMenu}) => {
   ), [])
 
   return (
-    <Layout.Content data-testid="layout-content-overlay" onClick={closeSideMenu}>
+    <Layout.Content data-testid="layout-content-overlay">
       <Router history={history}>
         <Switch>
           {configuration.plugins?.filter(hasRoute).map(routerMapper)}
@@ -31,8 +24,6 @@ export const LayoutCenter: React.FC<LayoutCenterProps> = ({closeSideMenu}) => {
     </Layout.Content>
   )
 }
-
-LayoutCenter.propTypes = layoutCenterProps
 
 const CenterPluginManager: React.FC<Plugin> = (plugin) => {
   return (
