+++ packages/fe-container/src/components/side-menu/SideMenu.less
@@ -1,7 +1,16 @@
+@import "../../variables";
+
 .sideMenu_menu {
-  height: calc(100vh - 64px);
+  height: calc(100vh - @top-bar-height);
   padding-top: 24px;
   border-top: 1px solid #d9d9d9;
+}
+
+.sideMenu_overlay {
+  top: @top-bar-height !important;
+  width: 256px;
+  position: fixed;
+  left: 0 !important;
 }
 
 .sideMenu_divider {
