+++ packages/fe-container/src/components/top-bar/TopBar.tsx
@@ -1,8 +1,12 @@
 import React, {useContext, useState} from 'react'
 import {FormattedMessage} from 'react-intl'
+import {Dropdown} from 'antd'
+
+import {MenuOpenedContext} from '../../contexts/MenuOpened.context'
+import {SideMenu} from '../side-menu/SideMenu'
+import {ConfigurationContext} from '../../contexts/Configuration.context'
 
 import './TopBar.less'
-import {MenuOpenedContext} from '../../contexts/MenuOpened.context'
 
 export const TopBar: React.FC = () => {
   return (
@@ -24,12 +28,21 @@ const BurgerIcon: React.FC = () => {
     setMenuOpened(!isMenuOpened)
   }
 
+  const configuration = useContext(ConfigurationContext)
+
+  const sideMenu = <SideMenu plugins={configuration?.plugins}/>
+
   return (
-    <label htmlFor="check" onClick={manageToggle}>
-      <input checked={isMenuOpened} data-testid="top-bar-side-menu-toggle" readOnly={true} type="checkbox"/>
-      <span/>
-      <span/>
-      <span/>
-    </label>
+    <Dropdown
+      onVisibleChange={manageToggle} overlay={sideMenu} overlayClassName="sideMenu_overlay"
+      trigger={['click']} visible={isMenuOpened}
+    >
+      <label htmlFor="check">
+        <input checked={isMenuOpened} data-testid="top-bar-side-menu-toggle" readOnly={true} type="checkbox"/>
+        <span/>
+        <span/>
+        <span/>
+      </label>
+    </Dropdown>
   )
 }
