+++ packages/fe-container/src/containers/launcher/Launcher.less
@@ -1,8 +1,10 @@
+@import "../../variables";
+
 .launcher_skeleton {
   width: 100% !important;
 
   span {
-    height: 64px !important;
+    height: @top-bar-height !important;
   }
 }
