+++ packages/fe-container/src/variables.less
@@ -1,0 +1,1 @@
+@top-bar-height: 64px;
